Changed node model format and added user authentication methods

+++ builder/models.py
@@ -8,10 +8,12 @@ class NodesManager(models.Manager):
 class Nodes(models.Model):
 	objects = NodesManager()
 
-	#lat = models.FloatField()
-	#lng = models.FloatField()
-	lat = models.DecimalField(max_digits=16, decimal_places=14)
-	lng = models.DecimalField(max_digits=16, decimal_places=14)
+	lat = models.FloatField()
+	lng = models.FloatField()
+	#lat = models.DecimalField(max_digits=16, decimal_places=14)
+	#lng = models.DecimalField(max_digits=16, decimal_places=14)
+	#lat = models.DecimalField(max_digits=8, decimal_places=6)
+	#lng = models.DecimalField(max_digits=8, decimal_places=6)
 	label = models.CharField(max_length=100, null=True, unique=True)
 
 	def __unicode__(self):
